+++ test/src/concurrent-secret-test.js
@@ -1,6 +1,6 @@
 import { randomInt } from 'node:crypto';
 
-import { startServer } from '@aller/google-cloud-secret/fake-server/fake-secret-manager-server';
+import { startServer } from '@aller/google-cloud-secret/emulator';
 import secretManager from '@google-cloud/secret-manager';
 import * as grpc from '@grpc/grpc-js';
 import * as ck from 'chronokinesis';